requirements_v2.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
blosc2==3.6.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
efficientnet_pytorch==0.7.1
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.5
huggingface-hub==0.33.4
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lightning==2.5.2
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
ndindex==1.10.0
networkx==3.4.2
numexpr==2.11.0
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-louvain==0.16
pytorch-lightning==2.5.2
pytubefix==9.5.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
safetensors==0.5.3
scipy==1.15.3
segment-anything==1.0
sentry-sdk==2.33.2
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tables==3.10.1
termcolor==3.1.0
timm==1.0.18
torch-geometric==2.6.1
torchmetrics==1.8.0
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.0
yarl==1.20.1
